Remove recursive imports.
